This commit is for NWA-75 validate the image URL and web site URL in "Edit Persional Info" page.
Added a check of length for Image Host and Website.  If their length exceeds 200 characters,
we stop accepting input characters and give a red warning below the field.
The warning goes away once you remove a character from input.

For both fields, added check whether they are valid URLs. If they are not, a red warning
is given below and Save button is disabled.