This is resolved.  Code is on Dev.
That wasn't a bug of NDEx Web App, but rather an issue with Cytoscape.js.
A bug for Cytoscape.js has been created https://github.com/cytoscape/cytoscape.js/issues/2152.

NDEx WebApp correctly processes continuous mapping of colors and builds CSS code.
Function mapData() of Cytoscape.js, however, fails to correctly map a color in case color is the same for the entire range.  So, for  selector
selector='edge[score > 2.5002426779747005][score < 4.4712]'
our code correctly created mapping of line color to be
line-color='mapData(score,2.5002426779747005,4.4712,rgb(255,255,0),rgb(255,255,0))'

With the two last arguments of mapData being the same rgb(255,255,0) (yellow), however,
cytoscape.js incorrectly set the color-line to gray.

A workaround for this case is to check if the range of values we map to is the same, and if it is,
then do not use mapData function, but use the value instead (since it is static anyways).