This commit is for NWA-199 Add "Set Sample" button to network view page.
Added Setting Network Sample via UUID of sample network.  It is added
as a menu item to a drop up menu on Network page, More button.

This menu item is greyed out in case network has a layout and no more than 12000 edges;
or it has less than 100 edges.

getNetworkAndDisplay() function in network-controller.js is augmented with a new condition:
if network has a layot and no more than 12000 edges, it is retrieved from
the server completely.