This prevents NWA-275 Failed API call in Web app.
The server log attached to NWA-275 reveals that API that fails with user null error
is /v2/user which corresponds to Web App call ndexService.getUserByUserNameV2:

        .then(
            ndexService.getUserByUserNameV2(editor.doiInfo.user.username,
                function(data) {
                    var userId = (data && data.externalId) ? data.externalId : null;
                    if (userId) {
                        editor.doiInfo.user.email = data.emailAddress;
                    }
                    else {
                        $scope.forgot.errorMsg = 'Unable to get User Id for user ' + $scope.forgot.accountName +
                            ' and request password reset.';
                    }
                },
                function(error){
                    editor.errors.push(error);
                })
        );

The only reason the above code is called is to get user's email address.  However, email address of a logged in user
is available in window.currentNdexUser.emailAddress.  So the above code is redundant.

The code has been modified to use email address from window.currentNdexUser.emailAddress and not call ndexService.getUserByUserNameV2
that sometimes results in the error described in NWA-275.

On the other hand, it is unclear why editor.doiInfo.user.username sometimes is null.