This resolves NWA-311 Error when creating new account.
When a new user was created and verification e-mail wasn't sent, we redundantly called
ndexService.authenticateUserV2 with basicAuthSuccessHandler as callback.

basicAuthSuccessHandler uses $scope.signIn.password (which was null if
basicAuthSuccessHandler is called after creating a new account), so
ndexUtility.setUserInfo set the password of the new user to null.
This null-password was used in all calls to the server, resulting in
authentication failures.

The code has been refactored not to call ndexService.authenticateUserV2 user
after creating a new user (this is redundant), and call basicAuthSuccessHandler(data)
directly with initialized data structure and valid $scope.signIn.password value.