Another commit for NWA-323 Revamp of sign in/sign up interface.
All Sign out activities are now carried out in  main.js/signOutHandler().  The main
change in this function is $window.location.href = '#/‘; which is the same as
$location.path('/');, but causes page reload.  This is done in order to eliminate
incorrect behavior we observed over the past few days when user Signs Up with Google,
then signs out, but the page shown is still My Account, until browser is refreshed.

Main.js/googleUserHandler() now correctly handles sign up errors.

Added new $scope.getClassForMyTasksTab to my-account-controller.js.  It returns
‘enabled’ or ‘disabled’ class to the caller to show My Tasks tab as enabled or disabled.
My Tasks & Notifications tab is disabled (shown in grey) in case there are no tasks or notifications
for the logged in user.

In myAccount.html, changed color of Refresh Glyphs in My Networks and My Tasks tabs to be the same
 when refresh is in progress.