These changes are for NWA-282 New Home Page for NDEx.
IN the top navigation bar, the NDEx server version color is set to black
to hide it until it is retrieved from the server.  Once it is retrieved, the
main.ndexServerVersion variable is initialized and its' color changes to blue.
This is done in order to hide the uninitialized
'v{{main.ndexServerVersion}}' angular expression shown in the top menu while
server version is being retrieved.

The fillInFeaturedContentChannel() function in main.js is simplified to retrieve
all featured objects (groups and users) in one loop.