This commit is for NWA-282 New Home Page for NDEx.
Fixed footer of landing page.  Before, if you shrink browser horizontally and
increase its' width back again, the browser would show white blank part in place
of the footer that was invisible.

Fixed alignment of FOLLOW US to be always right-aligned, and aligned with
Linked In image on the right.