This commit is for NWA-337 WebApp locking issue.
In app.js we redefine $location.path service to take second argument 'reload'.
If 'reload' is set to false, then the re-defined $location.path remembers the current
location, calls the original $location.path() and as soon as location is changed
($locationChangeSuccess event is triggered), changes the $route.current to previous
route; thus, the URL i the browser is changed, but page reload is not called.

The re-defined $location.path with the second argument set to false is called from 2 places in myAccountController: one is when we flip the pages of My Networks (), and another one is when page reload is detected.

The problem was that when page reload was called from My Account page,  $location.path() was called but $locationChangeSuccess defined in $location.path() was not called (it was already called before that).
Thus, next time $location.path() was called, it used to modify URL but show the previous route.

To fix this behavior, if we detect that reload was hit on My Account page, we DO NOT remember current route
for going there at the next $locationChangeSuccess event.