This commit resolves for NWA-358 Display error in search result page.
There is some conflict between Angular UI Grid and ngAnimate that we
inject into app.js in order to created sliding effect of Featured content.
Removing ngAnimate restores the previous correct behavior of
on the Search page when switching tabs.

If ngAnimate is iniected, we need to add timed refresh of the tables
 in order to make Switching tabs on the Search page works correctly.

 Sinice we do not know what other side effects ngAnimate introduces,
 we decided to remove it.