This commit is for NWA-397 Quick Edit feature for Network page.
Added Quick Edit feature for
1) Network Name,
2) Description,
3) Reference, and
4) Version

on Network page.

Selecting Quick Edit image (pencil) next to one of the above properties will
display a modal where this property can be edited.

After editing is done, modal is removed and only the edited network
attribute/property is refreshed in case its' original value was changed.
 The entire Network page doesn't get reloaded.

Also, fixed the logic of networkService.js/factory.setNetworkProperty()
to find property in NetworkSummary.properties objects with the specified
property name.  If this property is found, its value is updated.
If it is not found, it is created and added to the list of network summary
properties.