This commit is for NWA-397 Quick Edit feature for Network page.
Added Quick Edit feature for
1) Visibility

on Network page.

This feature allows to change network visibility (Private or Public),
make network searchable (with the option of full index) or not searchable,
and make public network showcased.

Quick Edit for visibility is disabled if either of Network Name, Decsription or
Version is missing.