This commit is for NWA-397 Quick Edit feature for Network page.
Changed the styles of enabled and disabled Edit Icon.  Disabled icon is a little smaller (0.9em)
than enabled.

Added getClassForQuickEdit() function to controller.  This function
returns class for enabled or disabled edit icon based on whether network
is editable or not. For Visibility parameter, it additionally checks
whether name, description and version are present.

Moved logic of checking if edit icons are enabled and using the class to controller.
Same with execution editNetworkSummaryProperty() and editVisibility():
we now always call them and the function checks whether attribute
is editable and whether to proceed or just return.