This resolves NWA-403 Full indexing checkbox should be disabled if user is not owner of a network.
Modified edit-network-properties-controller to initialize $scope.disabledVisibilityTooltip and
$scope.disabledFullIndexTooltip in case user is not admin of the network (s)he edits.

In mainNetworkProperties view, for Full Index we check if user is admin.  If no, then we disabled
this Full Index checkbox and pop a tooltip disabledFullIndexTooltip when user hovers over it.

Fixed tooltip for disabled Visibility - before if user couldn't change visibility because (s)he
wasn't an admin of this network, we showed a "slow" title description; now we show a "fast" tooltip description.

Fixed "used before initialized warning" by placing
var networkId = networkExternalId;
at the top of the controller.