This commit resolves NWA-400 Network lost background when switching back from table view
The default background color of  Cytoscape Graph is now always set to rgba(0,0,0,0)’;

In initCyGraphFromCyjsComponents() we create Canvas and set it under the Cytoscape graph.
To set the color of this canvas, we try to get it background color from our network in cx:
If it is present, it is set as background color (fillStyle)
If there is no background color in the network, we use the default rgba(0,0,0,0).

Important: in order to keep  background color of Canvas when switching from Graph view and back, and resizing the window,
 we need to catch the render and cyCanvas.resize events, re-set fillStyle and re-fill:

cy.on("render cyCanvas.resize", function() {
    ctx.fillStyle = cxBGColor;
    ctx.fillRect(0, 0, canvas.width, canvas.height);
});

For more info and sample code:
https://github.com/classcraft/cytoscape.js-canvas
https://github.com/classcraft/cytoscape.js-canvas/blob/master/demo.html