This resolves NWA-166 web-app send unnecessary calls to the server in myAccount page.
Modified code to only retrieve User Permissions for sent and received requests if
sent/received request(s) for the current user actually exist.

The same is for the Groups: we sent request to the server to get Group(s) only if the current
user is a member of group(s).